Move the "beginning withdrawal address" back to config.ini (#45)
We only accept the "beginning withdrawal address" if it can be generated by the configured "master public key". So even if we got hacked and the "beginning withdrawal address" got replaced by a wrong address, `bitcoin-dca` would not accept that invalid withdrawal address.

Moving the "beginning withdrawal address" back to the `config.ini` so that it can be easily modified. This would be handy when we want to reserve a few more addresses for other purposes